Update patch set 1
Patch Set 1: -Code-Review

It might be worth reverting this patch.  We can then operate on getting the needed structure in place based on Brady's content in 7433.  The structure there will not contribute to the release docs but we can push a patch to correct the structure to accomodate it.

Patch-set: 1
Reviewer: Gerrit User 163 <163@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0