Update patch set 2
Patch Set 2: Code-Review+1

> This is actually a dummy way to do it. I'll try a second patch set
 > to preserve the git history

We talked with Brady about this and decided that it is not worth it to bring every commit from functest individually. 
The history will be kept in functest. I'm adding the first and last relevant functest commits so the whole testcase history is between those.

Patch-set: 2
Label: Code-Review=+1