Update patch set 2
Patch Set 2: Code-Review-1

(12 comments)

Some comments inside. Probably they were already in the code, as it seems that you just moved from one place to another, but IMHO they are nice improvements.

Also, as a general consideration, I think that we should try to avoid "defaulting to PASS" when initializing results. This does not need to be addressed in this patch, but since it is one of the first patches I review, it made me think that it *might* lead to false positives, which is a dangerous thing.

The overall structure looks good to me, at least it is a big improvement over having one big file.

Patch-set: 2
Label: Code-Review=-1