Update patch set 5
Patch Set 5:

> (2 comments)

thread.start_new_thread(someFunc, ()) this approach has a limitation in that once the start_new_thread() function is called, it is not possible to find out when the thread has finished. Instead of waiting for 100 sec it makes more sense to wait for this thread to finish where(using join) we figure out flows have updated.  

Python also provides the threading module which implements a layer on top of the thread module. The threading module provides, among other things.
This link below explains its usage quite well.
http://softwareramblings.com/2008/06/running-functions-as-threads-in-python.html

Patch-set: 5