Update patch set 5
Patch Set 5:

> > (2 comments)
 > 
 > thread.start_new_thread(someFunc, ()) this approach has a
 > limitation in that once the start_new_thread() function is called,
 > it is not possible to find out when the thread has finished.
 > Instead of waiting for 100 sec it makes more sense to wait for this
 > thread to finish where(using join) we figure out flows have
 > updated.
 > 
 > Python also provides the threading module which implements a layer
 > on top of the thread module. The threading module provides, among
 > other things.
 > This link below explains its usage quite well.
 > http://softwareramblings.com/2008/06/running-functions-as-threads-in-python.html

Agreed and it's a nice addition. But I think we don't need to define a new class and a wrapper function to do this. You can write

t1 = threading.Thread(target=test_utils.capture_time_log,
                      args=compute_clients)
t1.start()

inline. It's more explicit, easier to understand and you don't need to go back and forth between different modules for such a small task.
What I want to say is that abstraction has a cost and when we can be more explicit we should.

Patch-set: 5
Label: Verified=0