Modify printouts of SFC test case to make them more readable
Created a new utility class: TerminalColors. This class can handle the
requirements on color codes, and be easily extended to support more editing
formats (e.g. background color, bold, etc.).

For the moment, only light_red and light_green (the only ones previously used)
have been added to this class. However, it should be straightforward to add new
ones.

Also, some small changes to the test printouts have been done, with the goal
of making the more uniform and avoid using inline string assigments when
the result of the assignment is already known.

Change-Id: I5474247da9a5fabf97024105c609832b74562f49
Signed-off-by: Juan Vidal <juan.vidal.allende@ericsson.com>