Update patch set 4
Patch Set 4:

Agree with Manuel. Should this be a class with static methods or simply a utility function?

To me it looks simpler to have a utility like:

def foreground(text, color_name):
    color_codes = {....}
    dostuff...

Also let's not introduce a new module for this, rather we should move it in utils.

Patch-set: 4