Update patch set 1
Patch Set 1: Code-Review-1

I don't see the point of this. We use vxlan_tool as an external tool to setup a firewall and it does it. If it changes in a breaking way we would have the test failing and having to refactor. 

In fact the better way to do this is to pin the github url you added to a specific commit hash which defeats the purpose of this change.

Patch-set: 1
Reviewer: Gerrit User 5579 <5579@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1