Update patch set 4
Patch Set 4:

> I disagree with you here. I strongly prefer to have small, reusable
 > classes with clearly defined intents, rather than "utils"
 > collections with all kind of unrelated functions. These utilitiy
 > classes/packages tend to grow without much control, and they end up
 > being very difficult to reuse. Also, dependency management ends up
 > being a hell, as everything depends on utils... for very different
 > reasons.
 > 
 > For the current utils.py, my preference would be to split it into
 > different clases with stronger consistency (e.g. one for running
 > commands, one for vxlan_tool-related stuff, another for
 > time-related things, and so on).
 > 
 > Also, even when this TerminalColor class has just a couple of
 > functions for now, if we find it useful it could also be extended
 > to handle background colors, or even fancier effects, all related
 > to the same purpose: changing the terminal font rendering.
 > 
 > That's my generic approach to programming. However, this is a team
 > decission, so if you still prefer to move it to utils, I'll do
 > that.

That's a more philosophical discussion that we can take somewhere else (let's not explode the comment section in gerrit) :). In short my take on this is that any new abstraction, increase in complexity should be accurately and properly justified (not just "we may need it in the future"), otherwise what can be kept dead simple and explicit should be. 

To the point, my reasoning is that we are in a very active development phase where we are moving things around a lot and I would prefer not to bloat the repo with a new module for any new small feature / refactoring task.

My +1 remains. It's a team decision as you said

Patch-set: 4