Update patch set 4
Patch Set 4:

I don't think a Class is needed, since all the methods will be static anyway. It has no advantage over a terminal_color module with flat functions. There are cases for Classes and Objects but here it is just overkill imo and using @classmethod is bad style.

Patch-set: 4