Update patch set 1
Patch Set 1:

> > I don't see the point of this. We use vxlan_tool as an external
 > > tool to setup a firewall and it does it. If it changes in a
 > > breaking way we would have the test failing and having to
 > refactor.
 > >
 > > In fact the better way to do this is to pin the github url you
 > > added to a specific commit hash which defeats the purpose of this
 > > change.
 > 
 > ok. How about maintaining vxlan_tool.py in odl-sfc repo only and
 > transfer it to SF's runtime. Or do you any other idea in mind?
 > 
 > There are 2 scenarios here. Where VNF image has to be changed but
 > we want vxlan_tool.py in it or we may need latest vxlan_tool.py in
 > existing VNF image. Any of this requires changing image with
 > vxlan_tool.py in it.

The solution to both of these is to automate the image creation (with vagrant for example) and then pick the commit hash we want from the remote repo. I agree with you that we may have an issue if we want to change the image, but this is a task for the image deployment not the test runtime.

Patch-set: 1