Update patch set 1
Patch Set 1:

No I mean when someone wants to run the test with a new image he should create the image and upload to artifacts (manually / with a jenkins job...) and then point the configuration of the test to it.

The test logic should remain unchanged regardless of what image is running or which tools should be deployed in that image. That's not a responsibility for the testcase that's a responsibility for the system setup.

Take this example. If we used 10 different tools in the image instead of one would we be responsible to maintain and deploy all of them? If we used ubuntu and fedora images should the test have code to handle the different installation processes?

Patch-set: 1