Update patch set 4
Patch Set 4:

(2 comments)

I think its ok that the terminal_color code is in a class with static methods. The question is where to put it, I think its probably best for it to be in the functest repo so its usable by everybody.

Also, I would prefer the color strings to be some sort of enum accessible on the static class.

Patch-set: 4
Reviewer: Gerrit User 1499 <1499@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0