Update patch set 4
Patch Set 4:

I'm also ok to have this in the utils function. Objects should be created only when really necessary. I'm in favor of object oriented programming style, but an object with just a single function doesn't make much sense. An example of a good object could be for example an SF, with its methods create(), clean(), etc...

Patch-set: 4