Update patch set 1
Patch Set 1:

> We may also have to delete example-net, example-subnet,
 > example-router, example-sg etc.

This script is just a way to delete things when testing. Normally functest will delete everything automatically. Therefore, I just want to delete things which may disturb further testing (e.g. floatingips because we can end up depleting them or VMs because we run out of resources). The resources you mention are not duplicated if a test is run several times, so they don't disturb if not deleted and that is why I think they are out of the scope for this script.

Patch-set: 1