Update patch set 1
Patch Set 1:

Careful here. The second test is modifying the vnfds because we want to control where the SFs are deployed. In other words, test1 relies on nova to schedule the deployment whereas test2 decides where they are deployed based on the availability_zone field of the vnfd. That is why we had different vnfds per test. This change does not really affect us as long as test1 is executed before test2 but then the VMs of test3 will be deployed based on test2 decision. Not sure what to decide, maybe we can discuss by mail

Patch-set: 1