Update patch set 1
Patch Set 1:

I used a 'diff' command to check the vfnd files and they were exactly the same, that is the reason why I consider them duplicated.

Apparently the VNFD file allows you to specify the availability zone, but it was set to 'nova' as well in these files.

I completely agree that, if for any reason, the tests have different requirementes, they should have different configuration files, but in this case they were exacly the same.

Patch-set: 1