Update patch set 2
Patch Set 2:

> > > I don't like this approach.
 > > >
 > > > Utils.py should be just a utility library, the only ones
 > reading
 > > > configuration should be the test-cases themselves. They can ask
 > > > utils to do things on their behalf, but IMHO utils should not
 > > have
 > > > direct access to config.
 > >
 > > I don't get your objection. Please explain me what would be the
 > > benefit of not making use of CommonConfig inside our utilities
 > > library?
 > 
 > In my opinion,
 > 
 > > > I don't like this approach.
 > > >
 > > > Utils.py should be just a utility library, the only ones
 > reading
 > > > configuration should be the test-cases themselves. They can ask
 > > > utils to do things on their behalf, but IMHO utils should not
 > > have
 > > > direct access to config.
 > >
 > > I don't get your objection. Please explain me what would be the
 > > benefit of not making use of CommonConfig inside our utilities
 > > library?
 > 
 > I think that utils should be "stateless" and "dumb", meaning that
 > for every set of input parameters, you always get the same output,
 > not depending on external sources (i.e. configuration). That makes
 > the debugging (and unit testing, if we go that far) of utils much
 > simpler. Also, it reduces the number of crossed dependencies.
 > 
 > In my opinion, test cases should be the ones reading the config,
 > then they delegate on utils for some stuff (generic stuff common to
 > all testcases). Having the utils depend on config might introduce
 > subtle bugs.

The utils functions already are stateful. Take this function for
instance. Create_vnf depends on the state that the correct VNFD 
is already created which depends on the state that the correct 
VNFD files exist. This is true for almost all functions in utils.py. A misconfigured testcase will and should fail  anyway.

I won't argue this to death and I don't mind adding another 
argument to the function definition, but I think we are beating 
our selves here with overengineering an anti-pattern (utils 
module with random functions) to get a design feature that we 
already don't (and based on the nature of the project can't) 
have.

Patch-set: 2
Label: Verified=0