Update patch set 6
Patch Set 6:

Since we need to get this patch in for MS6 today, the pending comment, although is valid, is not that important. Juan will continue to work on this after the milestone, and he will address the comment then. Im going to go ahead and merge this.

Patch-set: 6
Label: Verified=0