Remove get_floating_ips()
get_floating_ips() is non-reusable function, with too much logic
into it. By using smaller functions, we can compose the same
functionality and build all tests upon a common set of utilities.

Using the new functions in functest to retrieve the nova ID for a VNF instance,
it is possible to use a generic solution at the problem of getting floating
ips deterministcally to the instances

Change-Id: Ic7dba908fa6bb343c177fe1a68322d3803ed1707
Signed-off-by: Juan Vidal <juan.vidal.allende@ericsson.com>