Reduce logging level of run_cmd
run_cmd does not have enough information to determine if something should
be logged with ERROR level. For example, you might want to repeat a command
until it succeeds, and you would expect it to fail the first times, so it
would not be an error in that context.

Let's keep the debug logging only, which should be informative for
troubleshooting.

Change-Id: I72bc40e58d19188b83ad289fe7fa2b4453d5ad56
Signed-off-by: Juan Vidal <juan.vidal.allende@ericsson.com>