Remove SFF_filter
After further investigation we found out that the classifier is deployed in all
computes. When deploying several tests, ODL starts to function very slowly and
sometimes the classifier takes a long time to be deployed in one compute. That
tricked us and made us think that it was not deployed everywhere

Change-Id: I474a05f77a6342be741d14ffbdb674ef7f970908
Signed-off-by: “Manuel Buil” <mbuil@suse.com>