Disable resource_finder_cache in controllers
Heat caching somehow messes our tests so we disable it before the tests
start.

For instance, if we do:

1. create example-net with NET_ID1
2. create VNFD1 with example-net
3. delete VNFD1
4. delete NET_ID1
5. create example-net with NET_ID2
6. create VNFD2 with example-net

heat has NET_ID1 cached as the id of example-net and tries to use that
instead of NET_ID2 to create VNFD2. This off course leads to a crash
since NET_ID1 does not exist anymore.

In the future we should investigate further why heat cache does not play
well for us (probably bc of tacker).

Change-Id: Iac07d78641e2d1ff0c2d67e46752d551833f19fd
Signed-off-by: George Paraskevopoulos <geopar@intracom-telecom.com>