Bug fix in CI: security groups
CI tests create a lot of security groups with the name default. Our SFs are
assigned to a default security group but there are so many that the result is
not predictable and most likely wrong. We add it now to example-sg sec. group

Change-Id: I120765a33e55e47b78ef3c8664c82fb147ce9df0
Signed-off-by: Manuel Buil <mbuil@suse.com>
(cherry picked from commit 6282c5cce71ae994aaece374375234cf207992f8)