Bug fix in CI: security groups not working
CI tests create a lot of security groups with the name default.
Our SFs are assigned to a default security group but there are so
many that the result is not predictable and most likely wrong.
We add it now to example-sg sec. group

Change-Id: I64984aa7b6114747c5bab5fb2e4b8d8d32865bb0
Signed-off-by: Manuel Buil <mbuil@suse.com>
(cherry picked from commit 3c6e0efc2f4575ec2cf2e1304bdc1b66cdc08b06)