Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I find this solution very convoluted and error prone. IMHO it adds uneeded complexity to the code.

I suggest to create an image for ARM architectures and store it in artifacts. Then we can just choose the correct image to download and avoid modifying the image existing at runtime.

Juan has created a Vagrantfile that bootstraps the image we use so you could create the ARM image based on that:
https://github.com/jvidalallende/debootstrap-test/tree/opnfv-sfc

Patch-set: 1
Label: Code-Review=-1