Adapt the testcases to multiple installers
Right now the testcases are dependent to only one installer.
So when we are trying to run the testcases to multiple installers
the tests are failing.

This patch make the testcases independent to the installer that they
run.

The testcases are running until the point which they hit the tacker
client commands.They are failing in the tacker client commands
because we need the new library for the upstream tacker API.

The testaces have already tested to two installers (Fuel,Apex-ha).

Also this patch contains the changes which are required for the
adaptation to the new functest.

For more information see the Jira tickets [0],[1],[2].

[0]: https://jira.opnfv.org/browse/SFC-100
[1]: https://jira.opnfv.org/browse/SFC-101
[2]: https://jira.opnfv.org/browse/SFC-102

Change-Id: Id1e5d5c94a65ab8bdea9584fa833bfa0cdec6632
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>