[Bug Fix] Delete vnffgs in a reverse order
JIRA: SFC-106

In sfc_two_chain testcase we need to delete first the blue
and then the red vnffg because if we do it the other way around
the tacker produces a warning that the flow classifier of the vnffg red
is in use and cannot delete the chain.

The aforementioned warning is produced because during the execution
of the two_chain testcase we reassing the classifier from the red vnffg
to the blue vnffg. So if we try to delete first the red vnffg we get the
warning because the classifier is assigned and in use from the blue
vnffg.

Change-Id: I3cb7bef4b9b9e75eef4961fe7f63cbc1755130be
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>