Update patch set 8
Patch Set 8: Code-Review-1

I would prefer if we don't merge this patch. We can leave the code as it is or we will need to change it again when we fix the other problem. We should definitely add it in the release notes

Patch-set: 8
Reviewer: Gerrit User 2091 <2091@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1