Update patch set 8
Patch Set 8:

> I would prefer if we don't merge this patch. We can leave the code
 > as it is or we will need to change it again when we fix the other
 > problem. We should definitely add it in the release notes

If we do not merge this patch and the guys in releng merge the other patch.we are not gonna be compatible with the ovs-logger.

If we do not merge this patch and also we do not merge the releng patch we are gonna hit the error of index out of bound that we were hitting in the past.

Patch-set: 8
Reviewer: Gerrit User 6745 <6745@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0