Update patch set 3
Patch Set 3:

> Hi Manuel, the links in commit message refers to pylint rules in
 > question.

Hi Juha, thanks for the explanation but I still don't understand, sorry :S.

http://pylint-messages.wikidot.com/messages:w0702 says we should define an Exception type instead of not adding anything (As it was before) or adding just Exception as this patch is doing. As far as I see (and I might be wrong), the result is the same in both implementations and as the link says "Using a catch-all Exception instance defeats the purpose of knowing the type of error that occured"

Regarding PEP-0282 I agree to it but the only thing this patch does related to logging is moving the import logging to line 11 instead of line 18. And changing the syntax to print logs from a compatible syntax with 2.7 and 3.x to an exclusive syntax for 2.7. I also don't see how this change relates to PEP-0282

The env.get() change I understand :)

Patch-set: 3