DO NOT MERGE - Support sfp redirect action on acl flow check
Starting Oxygen, ODL networking-sfc translator supports symmetric
paths and maps flow classifiers to ACLs with SFP redirect action
instead of RSP redirect action. A SFP may have two RSPs associated
if the path is symmetric, one otherwise. One of the two RSPs will
be flagged as the reverse path and the other is the forward path.
Traffic from the logical-source-port will be classified to the forward
path, while traffic from the logical-destination-port will be
classified to the reverse path.

This patch updates the ACL flow check logic to account for the above.

Change-Id: I2d8b9e6c2a1033d469668db29b9c18f525e89370
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>