Update patch set 3
Patch Set 3: Code-Review-1

Yes Functest (both OpenStack/Kubernetes) has already been updated. 
But this patch has to be updated (xtesting should be listed in your requirements.txt even if it would be installed indirectly by Functest)

We detected an issue regarding our conf. Functest is calling your testcase as a BashFeature but it was written as a TestCase.

Then it has to depend on xtesting.

Yes

Patch-set: 3
Label: Code-Review=-1