Update patch set 23
Patch Set 23:

> did you ever run the testcases from the run_sfc_test script ?
 > 
 > Because I am not seeing any changes to that script and I am wonder
 > if you should make some changes to that script too.

Yes, I use the run_sfc_test script for the execution of test scenarios, adapting the config file if I want to add/remove any TC for the execution order.
I have adapted the code of each TC so that each TC to be able to run either via run_sfc_test script or stand alone. 
Due to this, no update on the run_sfc_test was needed.

Patch-set: 23
Label: Verified=0