Create unit tests for 'cleanup.py' module
This is the first attempt to add unit testing capabilities
to OPNFV/SFC project.

Also we are following the contraints.txt logic which
helps in the direction of simulating a test env which
will be as much as close to the real env that the SFC
testcases are running.

Jira: SFC-127
Jira: SFC-129

Change-Id: I938d28f9f85614cdefa163949c4ec9f06f5d27b4
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>