Improve logging and unify how to read from yaml
vnffgd and vnfd are being parsed in different ways, we should parse them
in the same way. We could do it directly with safe_load, however, the returned
object is not easy to read. If we use read() instead, the resturned object is
easy to read/log. Therefore, this patch combines both options

Change-Id: I3416933801569a6f57962cd9906726229923136d
Signed-off-by: Manuel Buil <mbuil@suse.com>