Update patch set 2
Patch Set 2: Code-Review-1

That is a great Idea but to be honetst I do not agree with the implementation.

IMHO I would like to have a default variable value setted to True for cleanup.

And if we want to disable it and to make it False then the user can actually pass that ass option when it executes the run_sfc_tests.py  script like this:

run_sfc_tests --cleanup False

Patch-set: 2
Label: Code-Review=-1