Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

One extra thing which could be done in this patch is removing the sys.exit(-1) that we have in the parent function, for example: https://github.com/opnfv/sfc/blob/master/sfc/tests/functest/sfc_parent_function.py#L263 if there is an error, then an Exception should be raised and catched by the testing framework. We used to exit through sys.exit(-1) instead because that was the only way at that time to avoid the clean up process (if we exit with sys.exit then the execution was not coming back to run_sfc_tests and thus cleanup was not called. I know, very ugly but worked :P )

Patch-set: 3
Label: Code-Review=-1