Update patch set 3
Patch Set 3: Code-Review+2

(1 comment)

we can merge this as it is now because @Manuel need this for Plugfest.But I think this is not flexible enough so we should replace in the future this functionality with a more dynamic one. So the user can add images directly from the command line when he executes the tests.

Patch-set: 3
Label: Code-Review=+2