Update patch set 3
Patch Set 3:

> (2 comments)
 > 
 > One extra thing which could be done in this patch is removing the
 > sys.exit(-1) that we have in the parent function, for example:
 > https://github.com/opnfv/sfc/blob/master/sfc/tests/functest/sfc_parent_function.py#L263
 > if there is an error, then an Exception should be raised and
 > catched by the testing framework. We used to exit through
 > sys.exit(-1) instead because that was the only way at that time to
 > avoid the clean up process (if we exit with sys.exit then the
 > execution was not coming back to run_sfc_tests and thus cleanup was
 > not called. I know, very ugly but worked :P )

I proposed a better solution for the sys.exit problem.
If we let it as it, we have the problem that you mentioned above.
If we remove it, the code execution is continuing and it probably will leaded to another error or endless loop.

Please check and comment the proposed solution at patchset4.

Patch-set: 3
Label: Verified=0