Update patch set 5
Patch Set 5:

> (1 comment)

Actually, I've found two ways to implement the part that initialize the clean up flag and checks whether the '--nocleanup' parameter has been given.
If you have another one , please let me know! :)

Alt1
-----
The SfcFunctest class has not __init__ method and the part that initialize the cleanup_flag to be part of run() method of this class <set patch : 1>.

Or a new method in the class could be created and we can call it through run() method.

Alt2
-----
If a __init__ method is created in order for the cleanup flag to be initiated then we have to add this instruction (i.e. super....__init__()). The reason is that if not, the attributes from super class are not recognized in the SfcFunctest class [e.g. self.details].
I faced this during testing, and added the super...__init__() in order to solve it.

Therefore, which approach do you believe that is better and less ugly :P?

Patch-set: 5
Label: Verified=0