Update patch set 1
Patch Set 1: Code-Review-1

But didn't we have to different yamls for this issue with tacker?

test-symmetric-vnffgd.yaml-queens is the one to use with the newest tacker version.

And the file modified in this commit is the one to use with the older tacker version, and imho, should stay as it is.

Patch-set: 1
Reviewer: Gerrit User 5925 <5925@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1