Update patch set 4
Patch Set 4:

> > (1 comment)
 > 
 > mmmmm, I don't understand what is the difference between [1] and
 > [2], it is the same python script being executed or? In both cases
 > it will access the script with __name__ = __main__. Do you
 > experience anything different?

hhmmm actually a correction in my aforementioned analysis! :)
Check the updated phrase from my previous comment, below:

"Therefore, if the run_sfc_tests.py is called via functest[0] or directly[1] the main() method is needed.
But, if we try to execute the run_sfc_tests file from python2.7/site-packages/... repository [2], we should first check if __name__ = __main__  and afterward to call the SfcFunctest.run() method directly. If there is not this checking , we could execute the sfc test scenarios only using the ways [0] and [1]."

Therefore, the difference that I can see [correct my if I am wrong] is,

-In cases of [0][1], the following instructions are executed:
if __name__ == "__main__":
    sys.exit(main())

-Otherwise in case [2], the only way to run the sfc test scenarios is to call the SfcFunctest.run() method from [3] itself. Of course before that we should check whether the __name__ is __main__.

[3] /usr/lib/python2.7/site-packages/sfc/tests/functest/run_sfc_tests.py

Patch-set: 4