Update patch set 4
Patch Set 4:

Hi Jaime, Is it the source or destination ports that are blocked in 2nd and 3rd check? From the description in the wiki it seems to block source ports.
If the destination ports are blocked then the interfaces(eth0 and eth1) have to be reversed.
I am not very clear, I am just trying to understand.

Patch-set: 4
Reviewer: Gerrit User 7321 <7321@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0