Update patch set 4
Patch Set 4:

> Hi Jaime, Is it the source or destination ports that are blocked in
 > 2nd and 3rd check? From the description in the wiki it seems to
 > block source ports.
 > If the destination ports are blocked then the interfaces(eth0 and
 > eth1) have to be reversed.
 > I am not very clear, I am just trying to understand.

In the symmetric test case, two instances of vxlan_tool are run. One that inputs eth0 and outputs eth1 for client->server, the other inputs eth1 and outputs eth0 for server->client traffic. Vxlan_tool is only able to block destination ports afaik. So depending on the direciton of traffic to be blocked, we use one or the other vxlan_tool instance.

Patch-set: 4
Label: Verified=0