Create os-odl-sfc_osm role
This change creates os-odl-sfc_osm role.
- copy the base scenario, os-odl-sfc as is
- add tasks, vars, defaults and templates from os-nosdn-osm role

It doesn't make any change to the copied os-odl-sfc role and
OSM deployment is done as post-deployment.

Verification of this patch will be done using a change in releng-xci
pointing to this commit/corresponding patch on Gerrit to verify it
works in order to help with the review.

Change-Id: If578e58fa96b7b83cbe5115b57a5059105bfefb8
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>