Update patch set 6
Patch Set 6: Verified+1 Code-Review+2

We decided to merge this beside the fact that jenkins gives -1. The -1 of jenkins is not related to the patch but because of another issue. This patch is tested locally and works so we are gonna merge it to save us from a lot of effort.

Patch-set: 6
Label: Verified=+1
Label: Code-Review=+2