Add ceilometer to the scenario
Due to the n-odl dependency, ceilometer must be part of the scenario.

There is a chicken-egg problem with this patch. There is a OSA patch
which will not get a +1 from jenkins if this patch does not get merged.
However, this patch will not get a +1 from jenkins unless OSA patch
gets merged. This is OSA's patch:

https://gerrit.opnfv.org/gerrit/#/c/62419/

Change-Id: I145ac35dcb94d40f6e21a1a49ae5a1e754434a10
Signed-off-by: Manuel Buil <mbuil@suse.com>