Update patch set 7
Patch Set 7:

I have uploaded more logs on the JIRA ticket.
It seems that the problem is solved using barbican keys instead of fernet keys. The scenario has been run end-to-end.
The -1 of jenkins-ci is not related to this patch, but due to document's a broken link.

Patch-set: 7
Label: Verified=0