Make the SFC TCs independently from installer
The SFC TCs are bound by the installers.
This means that the SFC TCs could not run on baremetal,
since some information are retrieved directly from
installer through deploy factory module.

A new yaml file is created in order for the important
information to be defined by user. This info could be used
as input in the TCs, since this is critical information for
execution of each testcase and it is related to the server itself.

JIRA: SFC-142

Change-Id: I051bb6406b8aa433c19e4df10396b4789448d42b
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>