Allow TCs to consume config info from user's file
The SFC TCs are bound by the installers.
This means that the SFC TCs could not run on server or VM
which all components (e.g. OS, ODL, etc) are deployed manually and
not by an installer, because some information are retrieved directly
from installer through deploy factory module.

A new yaml file is created by user in order for the important
information to be consumed by the respective test scenario during
test execution.

JIRA: SFC-142

Change-Id: I051bb6406b8aa433c19e4df10396b4789448d42b
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>